requirements.txt
# ============================================
# MLOps Standard Stack - Python Dependencies
# ============================================

# MLflow 및 실험 관리
mlflow>=2.10.0
boto3>=1.34.0

# TorchGeo - 지리공간 ML
torchgeo>=0.5.0
torch>=2.0.0
torchvision>=0.15.0

# 지리공간 데이터 처리
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.6.0
rtree>=1.0.0
fiona>=1.9.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0

# 유틸리티
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# 이미지 처리
pillow>=10.0.0
opencv-python>=4.8.0

# API 서빙 (선택적)
fastapi>=0.100.0
uvicorn>=0.23.0
